refactors for text/dpo evaluation
